new hybrid system (generic for all raytracing tools)

+++ orangecontrib/shadow4/util/shadow4_util.py
@@ -303,8 +303,8 @@ try:
                 super(ShadowPlot.InfoBoxWidget, self).__init__()
 
                 info_box_inner=gui.widgetBox(self, "Info")
-                info_box_inner.setFixedHeight(518*y_scale_factor)
-                info_box_inner.setFixedWidth(230*x_scale_factor)
+                info_box_inner.setFixedHeight(int(518*y_scale_factor))
+                info_box_inner.setFixedWidth(int(230*x_scale_factor))
 
                 self.flux_box = gui.widgetBox(info_box_inner, "", addSpace=False, orientation="horizontal")
                 self.flux     = gui.lineEdit(self.flux_box, self, "flux_field", "\u03a6 [ph/s/0.1%BW]", tooltip="Flux", labelWidth=115, valueType=str, orientation="horizontal")
@@ -615,8 +615,8 @@ try:
                 self.plot_canvas = gui.imageWiew(parent=self)
 
                 self.plot_canvas.setColormap({"name":"temperature", "normalization":"linear", "autoscale":True, "vmin":0, "vmax":0, "colors":256})
-                self.plot_canvas.setMinimumWidth(590 * x_scale_factor)
-                self.plot_canvas.setMaximumWidth(590 * y_scale_factor)
+                self.plot_canvas.setMinimumWidth(int(590 * x_scale_factor))
+                self.plot_canvas.setMaximumWidth(int(590 * y_scale_factor))
 
                 self.info_box = ShadowPlot.InfoBoxWidget(x_scale_factor, y_scale_factor)
 
@@ -625,8 +625,8 @@ try:
                 layout.addWidget(self.info_box,    0, 1, 1, 1)
                 layout.addWidget(self.plot_canvas, 0, 0, 1, 1)
 
-                layout.setColumnMinimumWidth(0, 600*x_scale_factor)
-                layout.setColumnMinimumWidth(1, 230*x_scale_factor)
+                layout.setColumnMinimumWidth(0, int(600*x_scale_factor))
+                layout.setColumnMinimumWidth(1, int(230*x_scale_factor))
 
                 self.setLayout(layout)
